refactor: move core game logic to separate component
